fix(misc): Fix ExpressionProvider api usage as it got moved to the api package
