+++ gradle.properties
@@ -1,5 +1,5 @@
 fiatVersion=1.26.0
-korkVersion=7.85.0
+korkVersion=7.87.0
 kotlinVersion=1.4.0
 org.gradle.parallel=true
 spinnakerGradleVersion=8.10.0
