+++ orca-core/src/test/groovy/com/netflix/spinnaker/orca/pipeline/expressions/PipelineExpressionEvaluatorSpec.groovy
@@ -17,7 +17,7 @@
 package com.netflix.spinnaker.orca.pipeline.expressions
 
 import com.netflix.spinnaker.kork.expressions.ExpressionEvaluationSummary
-import com.netflix.spinnaker.kork.expressions.ExpressionFunctionProvider
+import com.netflix.spinnaker.kork.api.expressions.ExpressionFunctionProvider
 import com.netflix.spinnaker.orca.api.pipeline.models.ExecutionStatus
 import com.netflix.spinnaker.orca.api.pipeline.models.PipelineExecution
 import org.pf4j.PluginManager
